downloader.v1.bak/requirements.txt
requests
psycopg2-binary
redis
yt-dlp
prometheus-client
Pillow